SISC_Jamundi/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
folium
streamlit-folium
geopandas
shapely
openpyxl
python-dotenv
requests
